# Documentation/jidelnicek_OpenAPI Spec.yaml
...
openapi: 3.0.3
info:
  title: Jídelníček API
  description: |
    API for the Jídelníček meal planning application designed for outdoor enthusiasts and expedition leaders.
    
    ## Authentication
    All endpoints except authentication and public recipe viewing require a valid JWT token.
    Include the token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Authenticated requests: 1000/hour per user
    - Unauthenticated requests: 100/hour per IP
    - Export operations: 10/hour per user
    
    ## Units and Localization
    - All weights are stored internally in grams
    - All energy values are stored internally in calories
    - Unit conversion happens at the presentation layer based on user preferences
    - Supported languages: English (en), Czech (cs)
  version: 1.0.0
  contact:
    name: Jídelníček Support
    email: support@jidelnicek.cz
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.jidelnicek.cz/api/v1
    description: Production server
  - url: https://staging-api.jidelnicek.cz/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /recipes:
    get:
      tags:
        - Recipes
      summary: List user's recipes
      description: Retrieve all recipes created by the authenticated user
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: search
          in: query
          description: Search term for recipe names
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, name]
            default: updated_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Recipes
      summary: Create new recipe
      description: Create a new recipe with ingredients and nutritional calculation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
            example:
              name: "Trail Mix Energy Bars"
              description: "High-energy bars perfect for hiking"
              instructions: "Mix all ingredients, press into pan, bake at 180°C for 20 minutes"
              prep_time_minutes: 15
              cook_time_minutes: 20
              water_ml: 0
              servings: 8
              ingredients:
                - ingredient_id: "123e4567-e89b-12d3-a456-426614174000"
                  quantity_g: 200
                - ingredient_id: "223e4567-e89b-12d3-a456-426614174000"
                  quantity_g: 150
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Recipe limit reached (500 recipes per user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /recipes/search:
    get:
      tags:
        - Recipes
      summary: Search recipes
      description: Advanced search for recipes with filters
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [breakfast, lunch, dinner, snack]
        - name: dietary
          in: query
          schema:
            type: array
            items:
              type: string
        - name: min_calories
          in: query
          schema:
            type: integer
        - name: max_calories
          in: query
          schema:
            type: integer
        - name: max_time
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Ingredient endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    limitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input"
            message: "The provided data does not meet the required format"
            details:
              - field: "email"
                issue: "Invalid email format"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Invalid or expired authentication token"
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          description: Additional error details
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Current offset
        has_more:
          type: boolean
          description: Whether more items exist

    # Authentication schemas
...
    NutritionalValues:
      type: object
      required:
        - calories
        - proteins_g
        - carbohydrates_g
        - fats_g
      properties:
        calories:
          type: number
          minimum: 0
          description: Energy in calories (kcal)
        proteins_g:
          type: number
          minimum: 0
          description: Proteins in grams
        carbohydrates_g:
          type: number
          minimum: 0
          description: Total carbohydrates in grams
        fats_g:
          type: number
          minimum: 0
          description: Total fats in grams
        # Optional detailed breakdown
        sugars_g:
          type: number
          minimum: 0
          description: Sugars (subset of carbohydrates)
        saturated_fats_g:
          type: number
          minimum: 0
        trans_fats_g:
          type: number
          minimum: 0
        monounsaturated_fats_g:
          type: number
          minimum: 0
        polyunsaturated_fats_g:
          type: number
          minimum: 0
        cholesterol_mg:
          type: number
          minimum: 0
        fiber_g:
          type: number
          minimum: 0
        salt_g:
          type: number
          minimum: 0
        calcium_mg:
          type: number
          minimum: 0
        sodium_mg:
          type: number
          minimum: 0
        water_g:
          type: number
          minimum: 0
        phe_mg:
          type: number
          minimum: 0
          description: Phenylalanine for PKU tracking
        # Vitamins
        vitamin_a_ug:
          type: number
          minimum: 0
        vitamin_b1_mg:
          type: number
          minimum: 0
        vitamin_b2_mg:
          type: number
          minimum: 0
        vitamin_b3_mg:
          type: number
          minimum: 0
        vitamin_b5_mg:
          type: number
          minimum: 0
        vitamin_b6_mg:
          type: number
          minimum: 0
        vitamin_b7_ug:
          type: number
          minimum: 0
        vitamin_b9_ug:
          type: number
          minimum: 0
        vitamin_b12_ug:
          type: number
          minimum: 0
        vitamin_c_mg:
          type: number
          minimum: 0
        vitamin_d_ug:
          type: number
          minimum: 0
        vitamin_e_mg:
          type: number
          minimum: 0
        vitamin_k_ug:
          type: number
          minimum: 0

    # Recipe schemas
    RecipeIngredientInput:
      type: object
      required:
        - ingredient_id
        - quantity_g
      properties:
        ingredient_id:
          type: string
          format: uuid
        quantity_g:
          type: number
          minimum: 0.1
          maximum: 10000
          description: Quantity in grams
    RecipeIngredient:
      allOf:
        - $ref: '#/components/schemas/RecipeIngredientInput'
        - type: object
          properties:
            ingredient:
              $ref: '#/components/schemas/Ingredient'
    RecipeCreate:
      type: object
      required:
        - name
        - ingredients
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        instructions:
          type: string
          maxLength: 2000
        prep_time_minutes:
          type: integer
          minimum: 0
          maximum: 1440
        cook_time_minutes:
          type: integer
          minimum: 0
          maximum: 1440
          description: Time on stove
        water_ml:
          type: integer
          minimum: 0
          description: Cooking water required
          default: 0
        servings:
          type: integer
          minimum: 1
          maximum: 100
          default: 1
        ingredients:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/RecipeIngredientInput'
...
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        instructions:
          type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        water_ml:
          type: integer
        servings:
          type: integer
        is_public:
          type: boolean
        published_at:
          type: string
          format: date-time
        fork_count:
          type: integer
        forked_from_id:
          type: string
          format: uuid
        can_unpublish:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        current_version_id:
          type: string
          format: uuid
        version_number:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredient'
        nutrition:
          $ref: '#/components/schemas/NutritionalValues'
        images:
          type: array
          items:
            $ref: '#/components/schemas/RecipeImage'
        author:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
    RecipeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    RecipeImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_url:
          type: string
          format: uri
        file_size_bytes:
          type: integer
        uploaded_at:
          type: string
          format: date-time

    # Ingredient schemas
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Null for global ingredients
        name:
          type: string
        category:
          type: string
        is_global:
          type: boolean
        nutritional_values:
          $ref: '#/components/schemas/NutritionalValues'
        created_at:
          type: string
          format: date-time
...